Move typescript utils to higher in dir hierarchy
