+++ src/utils/typescript.ts
@@ -1,0 +1,3 @@
+export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
+
+export type classnamesValue = string | { [key: string]: string | boolean };
